refactor(napi): move checker batch strategy to orchestrator
